Regroup codes for readability
