+++ luda-adventure/src/app/game/render/mod.rs
@@ -1,0 +1,9 @@
+mod create_rendering_context;
+mod render_background;
+mod render_guide_icon;
+mod render_in_screen_object_list;
+
+pub use create_rendering_context::*;
+pub use render_background::*;
+pub use render_guide_icon::*;
+pub use render_in_screen_object_list::*;
