+++ luda-adventure/src/app/game/update/mod.rs
@@ -1,0 +1,5 @@
+mod handle_character_movement_on_key_event;
+mod predict_character_movement_if_needed;
+
+pub use handle_character_movement_on_key_event::*;
+pub use predict_character_movement_if_needed::*;
