+++ luda-adventure/src/app/game/update/predict_character_movement_if_needed.rs
@@ -1,4 +1,4 @@
-use super::{player_character::PlayerCharacter, *};
+use crate::app::game::*;
 use namui::prelude::*;
 
 impl Game {
@@ -15,5 +15,23 @@ impl Game {
                 positioner.predict_movement(collider, collision_box_list.as_ref());
             }
         }
+    }
+
+    fn get_collision_box_list_without_character_collision_box(
+        &self,
+        current_time: Time,
+    ) -> Vec<CollisionBox> {
+        self.ecs_app
+            .query_entities::<(&Collider, &Positioner)>()
+            .iter()
+            .filter_map(|(entity, (collider, positioner))| {
+                if entity.id() == self.player_entity_id {
+                    None
+                } else {
+                    let position = positioner.get_position(current_time);
+                    Some(collider.get_collision_box(position))
+                }
+            })
+            .collect::<Vec<_>>()
     }
 }
